requirements.txt
fastapi
uvicorn
python-dotenv
lighter-v2-python
openai
motor
gunicorn
